Rename sugar::refptr to sugar::ownership
