+++ ports/geckolib/binding_tools/regen.py
@@ -323,25 +323,25 @@ def build(objdir, target_name, debug, debugger, kind_name=None,
             flags.append("--blacklist-type")
             flags.append("{}Strong".format(ty))
             flags.append("--raw-line")
-            flags.append("pub type {0}Strong = ::sugar::refptr::Strong<{0}>;".format(ty))
+            flags.append("pub type {0}Strong = ::sugar::ownership::Strong<{0}>;".format(ty))
             flags.append("--blacklist-type")
             flags.append("{}Borrowed".format(ty))
             flags.append("--raw-line")
-            flags.append("pub type {0}Borrowed<'a> = ::sugar::refptr::Borrowed<'a, {0}>;".format(ty))
+            flags.append("pub type {0}Borrowed<'a> = ::sugar::ownership::Borrowed<'a, {0}>;".format(ty))
     if "servo_owned_types" in current_target:
         for ty in current_target["servo_owned_types"]:
-            flags.append("-blacklist-type")
+            flags.append("--blacklist-type")
             flags.append("{}Borrowed".format(ty))
-            flags.append("-raw-line")
+            flags.append("--raw-line")
             flags.append("pub type {0}Borrowed<'a> = &'a {0};".format(ty))
-            flags.append("-blacklist-type")
+            flags.append("--blacklist-type")
             flags.append("{}BorrowedMut".format(ty))
-            flags.append("-raw-line")
+            flags.append("--raw-line")
             flags.append("pub type {0}BorrowedMut<'a> = &'a mut {0};".format(ty))
-            flags.append("-blacklist-type")
+            flags.append("--blacklist-type")
             flags.append("{}Owned".format(ty))
-            flags.append("-raw-line")
-            flags.append("pub type {0}Owned = ::sugar::refptr::Owned<{0}>;".format(ty))
+            flags.append("--raw-line")
+            flags.append("pub type {0}Owned = ::sugar::ownership::Owned<{0}>;".format(ty))
     if "structs_types" in current_target:
         for ty in current_target["structs_types"]:
             ty_fragments = ty.split("::")
