+++ ports/geckolib/gecko_bindings/sugar/mod.rs
@@ -6,4 +6,4 @@ mod ns_css_shadow_array;
 mod ns_style_auto_array;
 pub mod ns_style_coord;
 mod ns_t_array;
-pub mod refptr;
+pub mod ownership;
